bookread/requirements.txt
asgiref
click
defusedxml
diff-match-patch
Django
django-debug-toolbar
django-import-export
django-js-asset
django-light
django-mptt
django-tablib
et-xmlfile
gunicorn
h11
lxml
MarkupPy
odfpy
openpyxl
Pillow
pip
pkg-resources
psycopg2
psycopg2-binary
python-slugify
pytz
PyYAML
setuptools
six
sqlparse
tablib
text-unidecode
typing-extensions
Unidecode
uvicorn
xlrd
xlwt
